requirements.txt
# Agents Workshop Dependencies

# LangChain Ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0

# Multi-Agent Frameworks
pyautogen>=0.2.0
crewai>=0.50.0

# Lightweight Agents
smolagents[litellm]>=1.0.0

# Google ADK (reference only)
google-adk>=1.0.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Search Tools
tavily-python>=0.3.0

# Data & Validation
pydantic>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0
httpx>=0.27.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Development Tools
rich>=13.0.0  # Pretty console output
